+++ apps/taxes/src/modules/taxes/tax-provider-webhook.ts
@@ -1,6 +1,8 @@
+import { SyncWebhookResponsesMap } from "@saleor/app-sdk/handlers/next";
 import { TaxBaseFragment } from "../../../generated/graphql";
 import { ChannelConfig } from "../channels-configuration/channels-config";
-import { ResponseTaxPayload } from "./types";
+
+export type ResponseTaxPayload = SyncWebhookResponsesMap["ORDER_CALCULATE_TAXES"];
 
 export interface ProviderWebhookService {
   calculateTaxes: (payload: TaxBaseFragment, channel: ChannelConfig) => Promise<ResponseTaxPayload>;
