--- apps/taxes/src/modules/taxes/types.ts
@@ -1,3 +1,0 @@
-import { SyncWebhookResponsesMap } from "@saleor/app-sdk/handlers/next";
-
-export type ResponseTaxPayload = SyncWebhookResponsesMap["ORDER_CALCULATE_TAXES"];
